@@ -20,16 +20,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'eployeeCrud'`, () => {
+  it(`should have as title 'takLiciousGoodies'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('eployeeCrud');
+    expect(app.title).toEqual('TakLiciousGoodies');
   });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('.content span')?.textContent).toContain('eployeeCrud app is running!');
+    expect(compiled.querySelector('.content span')?.textContent).toContain('TakLiciousGoodies app is running!');
   });
 });
